WNS-62: Script to delete users

See merge request freelance-website/back!20

+++ app/base/querysets.py
@@ -1,13 +1,14 @@
 from django.db.models import QuerySet
 from django.utils import timezone
 
+
 class BaseManagerQuerySet(QuerySet):
 
-  def delete(self):
-    return super(BaseManagerQuerySet, self).update(deleted_at=timezone.now(), is_deleted=True)
+    def delete(self, soft=True):
+        if soft:
+            return super(BaseManagerQuerySet, self).update(deleted_at=timezone.now(), is_deleted=True)
+        else:
+            return super(BaseManagerQuerySet, self).delete()
 
-  def hard_delete(self):
-    return super(BaseManagerQuerySet, self).delete()
-
-  def restore(self):
-    return super(BaseManagerQuerySet, self).update(deleted_at=None, is_deleted=False)
+    def restore(self):
+        return super(BaseManagerQuerySet, self).update(deleted_at=None, is_deleted=False)
